Fix misuse of gevent's StreamServer
StreamServer calls the handler we provided to it for every new incoming
connection, giving it the socket for that connection. That call is
performed in its own greenlet. Once that handler terminates,
StreamServer considers the socket as dealt with, and therefore ready
for closure and elimination. Yet, in gevent up to 1.0, the socket was
left untouched by StreamServer. Starting in gevent 1.1 StreamServer
closes the socket once the handler returns.

Our handler used to spawn another greenlet, inside of which it was
running an infinite loop that read from the socket and took care of
executing the requests. Once it spawned that greenlet, the handler
would return. This worked correctly in gevent 1.0 but breaks now, as
the socket will be closed immediately after the reading loop starts.

This causes the client to get into a loop of connection, immediate
disconnection and reattempt.