Handle isolate cleanup
Since the last version (see https://github.com/ioi/isolate/issues/26),
isolate fails on init if the box already exists. We didn't cleanup in
two situations: when the box was kept around, and when the worker is
terminated in the middle of an execution. The former is easy to fix,
the latter is not, so we also essentially revert to the previous
behaviour by always calling cleanup before init.