Logs in AWS are between <pre> tags
Since forever the logs in cmsAdminWebServer are shown in simple text
format. When something goes wrong a stacktrace is generated, which
always has many endlines and tabulation characters. In HTML this chars
are collapsed in a single space making the stacktraces very long lines.

The fix I propose is just to put the message inside a <pre> tag which
preserves all the spaces.