Make __future__ imports consistent and reproduce py3 builtins in py2
First, make sure *all* files have *all* four __future__ imports and that
they are in alphabetical order (will help in searches & replaces).

Then, use the future library (which I added to requirements.txt in the
version shipped by Ubuntu 16.04) to recreate in py2 the same builtins
that are present in py3. The `future.builtins.disabled` module contains
objects named like the ones that have been removed in py3 implemented
in a way that raises an error whenever they are accessed or used. The
`future.builtins` module contains object named like the ones that have
been added in py3 and tries to replicate their semantics. The two
imports have to happen in that order as some names occur in both.

These imports are no-ops in py3.