Handle dynamic imports using imp(ortlib)
The use of __import__ is discouraged and the reload builtin function
has disappeared in py3. The imp module provides a way to do the same
thing that is supported in both py2 and py3. It is not the ideal
solution, as it has already been deprecated and superseded by importlib
in py3, but it probably will do until we drop py2 support.