Don't leave open files around in the FileCacher test
Wrap every io.open(...).read() in a with block so that they get closed
as soon as we're done with them. This was printing warnings in py3.

Also leverage the fact that fail(...) raises an exception and thus it is
as if it implies a return statement, in order to remove unnecessary
explicit returns, else blocks, etc.