Replace d.iterkeys/values/items() method calls with iter...(d) function calls
In py2 .keys() returns a list and .iterkeys() returns an iterator. In
py3 there is only .keys() which returns a view, which is iterable. The
six module offers a iterkeys function which translates to iterkeys on
py2 and to the iterable of the keys view in py3. The same holds for
values and items.