Bump dependency versions to track Ubuntu 18.04
This is something we wanted to do anyway, as v1.4 (the upcoming
version) won't be released before Apr 2018 and can therefore be
designed to run on Ubuntu 18.04. Yet there is also an urgency which is
due to the fact that monkey-patching doesn't work with Gevent 1.1 and
Python 3.6. See: https://github.com/gevent/gevent/issues/903