Port CWS to Jinja2: simplify submission_ and user_test_row
They were replicating some logic that was defined in the DB class and
that could readily be accessed using get_status() (provided the result
was not None, which we can ensure by creating it if it doesn't exist).

Also, they were doing some weird stuff by putting some HTML in a
variable and then displaying later. There was no need for that.