Fix backdoor in py3
We were opening the UNIX socket for the backdoor (i.e., the file in
/var/local/run/cms) using the _socket module, which is the low-level
library of sockets, written in C and not designed for public use. I am
not sure why we were doing that, there were no comments. My guess is
that at the time we couldn't use the normal socket module due to
gevent's monkey patching. And maybe we thought that _socket gave access
to the unpatched socket module, not to a different thing...

This caused an issue in py3, because gevent didn't recognize the object
we were passing to the BackdoorServer as a socket, as it was missing the
accept method. (In py2 that method was implemented in C, in py3 it is
implemented in Python and hence only available in the high-level module)

Anyway, gevent's socket module now appears to support UNIX sockets and,
from what I can, see using it instead of _socket seems to work just
fine.