Improve retrieval of information from shared-mime-info
We use shared-mime-info to get the list of all MIME types, possible
aliases, their natural language descriptions and their icon names. We
were manually parsing all the relevant system files, of which we needed
to know the exact locations. The xdg library can do most of these tasks
for us. It doesn't provide information about the generic icon names
(yet?), which means we have to keep retrieving it by hand, but even
there it can help us figure out where the files are located.

This, in fact, removes our direct dependency on shared-mime-info, as we
have good fallbacks in place. And, anyways, it comes preinstalled by
default on all modern platforms.