Allow limiting of CPU time per service during RunTimeTest
This is useful to test the performance of ES, for example. If all
services are running at 100% CPU, one would need a lot of cores to
prove that moving operations to sharded service is helpful. Limiting
the CPU time to something that allows all services to run concurrently
within the capabilities of the physical CPU makes this more obvious.