Remove language selector from CWS form for handing in user test archives
Such form is shown only if the submission format is language-agnostic:
the language selector is thus useless. (Even more so, among the builtin
task types the only one with a language-agnostic submission format is
OutputOnly, which doesn't allow user tests, hence the whole form for
sending in user test archives is useless).