Option to limit feedback to avoid reverse engineering test data
Before, we offered many bits of information for each testcase back to
the contestant: high precision time and memory, killing signal.
Combined with showing information for all testcases, this gave the
possibility to discover a lot of information about the test data.

This patch introduce an option for the task to limit this feedback. By
default, the option is active (limited feedback) but to preserve the
old behavior for existing tasks the updater sets it to inactive.

The idea comes from the IOI, and the initial code from Amir Keivan
Mohtashami (see in particular [1] and [2]).

[1] https://github.com/akmohtashami/cms/commit/cecd9863
[2] https://github.com/akmohtashami/cms/commit/06b217a7