Add a WSGI middleware to serve files instead of having Tornado do it
The Tornado WSGI adapter buffers the whole response before starting to
send it to the client. This causes unexpected spikes in memory usage and
could be a vector for a denial-of-service attack. The Tornado devs have
no intention of fixing this, as their WSGI support is sub-standard by
design.

This commit implements file serving as a "native" WSGI application and
inserts it as a middleware into our stack. The Tornado handlers that
want to serve files have to issue an empty response with some custom
headers. These requests will be intercepted by the middleware that will
replace them with the actual files, in an efficient streaming fashion.
The middleware also supports conditional requests (based on the ETag
header) and partial requests (based on the Range header).

This also uses a mixin class to provide the file-serving methods to the
handlers rather than a function that would dynamically extend a given
class (thus avoiding magic and improving static analysis).

Unit tests are introduced too.