Avoid having FC's temp dir removed if a service is started twice
File cacher temporary directory was the same every time the service
was started; it is created when the service starts and removed when it
terminates. If service X is running and we started it again, the
second copy will terminate immediately due to port conflict, and
happily cleanup the original service temp dir for the FC.

With the change, the temp dir is unique per instance, so there is no
problem with a second instance coming up. There is a slightly larger
possibility of leaving trash behind, but only if a service is closed
abruptly.