Avoid infinite recursion when exiting a service connected to LS
Finalize had a logging call between the termination of the socket and
the marking of the client as disconnected; therefore the logging
module attempted to send the log to LS, but this incurred in an error
as the socket was close. The new error was then logged again, causing
the infinte recursion.

With the logging call after the client has been marked as
disconnected, execute_rpc for the log call returns early without
errors.