Fix bug if used memory is between 1000 and 1024 KiB, MiB, ...
If the memory used (or other values used as parameter of format_size)
ends in the range [1000, 1024) KiB/MiB/GiB... AWS and CWS will fail
rendering that size and the submission details wont be shown! This
because the number of digits of the number was supposed being at most
3, not true anymore since 1024 is used as the base.

Allowing an extra significant digit for those cases is the proposed solution.